wip: create resolveImportedNamespace module
