+++ src/utils/resolveToValueExternal.js
@@ -1,13 +1,15 @@
 const fs = require('fs');
 const docgen = require('react-docgen');
 const resolveExportDeclaration = require('react-docgen/dist/utils/resolveExportDeclaration').default
+import resolveNamespaceExternal, { resolveExternalNamespaceImport, resolveExternalImport } from './resolveImportedNamespace';
 import { Array as toArrayExternal } from './expressionTo';
+import { traverseShallow } from 'react-docgen/dist/utils/traverse';
 // TODO: replicate this in repo
-const traverseShallow = require('react-docgen/dist/utils/traverse').traverseShallow;
+// const traverseShallow = require('react-docgen/dist/utils/traverse').traverseShallow;
 const recast = require('recast');
 const types = require('recast').types;
 const babylon = require('react-docgen/dist/babylon').default;
-const resolveExternalFilepath = require('./resolveExternalFilepath');
+// const resolveExternalFilepath = require('./resolveExternalFilepath');
 const isPropTypesExpression = require('./isPropTypesExpression');
 
 const {
@@ -40,162 +42,6 @@ const {
 
 } = docgen.utils;
 
-
-// converts:
-//  `export default class foobar {}`
-// or
-//  `export defaut function foobar() {}`
-// by putting the expression on it's own line
-// and exporting default it's identifier
-// i.e. 
-//  class foobar {}
-//  export default foobar;
-function reassignExportDefaultExpression(ast) {
-  traverseShallow(ast, {
-    visitExportDefaultDeclaration(path) {
-      const node = path.node;
-      if (
-        !n.Identifier.check(node.declaration) &&
-        n.Identifier.check(node.declaration.id)
-      ) {
-        path.insertBefore(node.declaration);
-        path.get('declaration').replace(node.declaration.id);
-      }
-      return false;
-    }
-  });
-}
-
-function recastExternalFilepath(externalFilepath, baseFilepath) {
-  const filepath = resolveExternalFilepath(baseFilepath, externalFilepath);
-  const src = fs.readFileSync(filepath, 'utf-8');
-  const ast = recast.parse(src, { source: 'module', esprima: babylon });
-
-  reassignExportDefaultExpression(ast);
-
-  return {
-    ast,
-    filepath,
-  };
-}
-
-function resolveNamespaceExternal(externalFilepath, baseFilepath) {
-  const __EXPORTS__ = '__EXPORTS__';
-
-  const external = recastExternalFilepath(externalFilepath, baseFilepath);
-
-  const namespace = [];
-
-  let defaultModuleExports;
-
-  const addNamespaceExport = (key, valuePath) => {
-    namespace.push([key, valuePath]);
-  }
-
-  types.visit(external.ast, {
-    visitExpressionStatement(path) {
-      if (isExportsOrModuleAssignment(path)) {
-        const expressionPath = path.get('expression');
-        const valuePath = expressionPath.get('right');
-        const keys = getMembers(expressionPath.get('left')).reduce(
-          (memo, { path, computed, argumentsPath }) => {
-            if (path.node.name !== 'exports') {
-              memo.push(path.node.name);
-            }
-            return memo;
-          },
-          []
-        );
-        // module.exports = { ... }
-        if (!keys.length) {
-          defaultModuleExports = valuePath;
-        }
-        // module.exports.foo = ...;
-        else if (keys.length === 1) {
-          // namespace.push([keys[0], valuePath]);
-          addNamespaceExport(keys[0], valuePath);
-        }
-        // don't process module.exports.foo.bar... = ...;
-      }
-
-      this.traverse(path)
-    },
-    visitExportDefaultDeclaration(path) {
-      addNamespaceExport('default', path.get('declaration'));
-      this.traverse(path);
-    },
-    visitExportNamedDeclaration(path) {
-      path.get('specifiers').each(specifierPath => {
-        const identifierPath = specifierPath.node.id
-          ? specifierPath.get('id')
-          : specifierPath.get('local');
-
-        addNamespaceExport(identifierPath.node.name, identifierPath);
-      });
-
-      const declaration = path.get('declaration');
-
-      if (declaration.value) {
-        declaration.get('declarations').each(declaration => {
-          addNamespaceExport(getNameOrValue(declaration.get('id')), declaration.get('init'));
-        });
-      }
-
-      this.traverse(path);
-    },
-  });
-
-  const namespaceObjectExpression = (keyValuePaths) =>
-    b.objectExpression(
-      keyValuePaths.map(([key, valuePath]) =>
-        b.objectProperty(b.identifier(key), valuePath.node)
-      )
-    );
-
-  if (defaultModuleExports && n.ObjectExpression.check(defaultModuleExports.node)) {
-    // TODO: there is an ordering bug here.
-    // TODO: make sure that the order of definitions
-    // TODO: default/named order is right
-    defaultModuleExports.get('properties').each(
-      propertyPath => addNamespaceExport(propertyPath.get('key').node.name, propertyPath.get('value'))
-    );
-    addNamespaceExport('default', { node: namespaceObjectExpression(namespace) });
-  }
-
-  // inject the namespace valuepaths
-  external.ast.program.body.push(
-    b.variableDeclaration('const', [
-      b.variableDeclarator(
-        b.identifier(__EXPORTS__),
-        namespaceObjectExpression(namespace)
-      )
-    ])
-  );
-
-  // find the injected value and return it.
-  return resolveIdentifierNameToExternalValue(__EXPORTS__, external.ast, external.filepath);
-}
-
-function resolveExternalNamespaceImport(importDeclarationPath, baseFilepath) {
-  n.ImportDeclaration.assert(importDeclarationPath.node);
-  const source = importDeclarationPath.node.source;
-  n.Literal.assert(source);
-  return resolveNamespaceExternal(source.value, baseFilepath);
-}
-
-function resolveExternalImport(variable, importDeclarationPath, filepath) {
-  const external = resolveExternalNamespaceImport(importDeclarationPath, filepath);
-
-  const matched = external.path.get('properties').filter(
-    propertyPath => variable === getNameOrValue(propertyPath.get('key'))
-  );
-  const valuePath = matched[0].get('value');
-
-  if (n.Identifier.check(valuePath.node)) {
-    return resolveToValueExternal(valuePath, external);
-  }
-  return { ...external, path: valuePath };
-}
 
 // internal from docgen.utils.resolveToValue
 function buildMemberExpressionFromPattern(path) {
@@ -308,7 +154,7 @@ function findLastAssignedValueExternal(scope, name, { filepath }) {
  *
  * Else the path itself is returned.
  */
-function resolveToValueExternal(path, { filepath }) {
+export default function resolveToValueExternal(path, { filepath }) {
   const node = path.node;
   if (n.VariableDeclarator.check(node)) {
     if (node.init) {
@@ -381,29 +227,3 @@ function resolveToValueExternal(path, { filepath }) {
   }
   return { path, filepath };
 }
-
-
-
-function resolveIdentifierNameToExternalValue(name, ast, filepath) {
-  var result = false;
-  traverseShallow(ast, {
-    visitIdentifier(path) {
-      if (path.node.name === name) {
-        result = resolveToValueExternal(path, { filepath });
-        return false;
-      }
-      this.traverse(path);
-    },
-    visitVariableDeclarator(path) {
-      if (path.value.id.name === name) {
-        result = resolveToValueExternal(path, { filepath });
-        return false;
-      }
-      this.traverse(path);
-    },
-  });
-  return result;
-}
-
-
-module.exports = resolveToValueExternal;
